Updates database reset script
Refactors the database reset script to correctly clear the database, including enums and the migrations table.

Also, updates the `db:reset` script in `package.json` to include running the migrations before seeding the database.

Finally, updates the migration commands to use `npx` to ensure the correct version of `typeorm-ts-node-commonjs` is used.